+++ extensions/positron-r/amalthea/crates/harp/src/utils.rs
@@ -5,6 +5,8 @@
 //
 //
 
+use c2rust_bitfields::BitfieldStruct;
+
 use std::ffi::CStr;
 use std::ffi::CString;
 use std::os::raw::c_void;
@@ -34,6 +36,57 @@ extern "C" {
         symbol: SEXP,
         envir: SEXP,
     ) -> c_void;
+}
+
+#[derive(Copy, Clone, BitfieldStruct)]
+#[repr(C)]
+pub struct Sxpinfo {
+    #[bitfield(name = "rtype", ty = "libc::c_uint", bits = "0..=4")]
+    #[bitfield(name = "scalar", ty = "libc::c_uint", bits = "5..=5")]
+    #[bitfield(name = "obj", ty = "libc::c_uint", bits = "6..=6")]
+    #[bitfield(name = "alt", ty = "libc::c_uint", bits = "7..=7")]
+    #[bitfield(name = "gp", ty = "libc::c_uint", bits = "8..=23")]
+    #[bitfield(name = "mark", ty = "libc::c_uint", bits = "24..=24")]
+    #[bitfield(name = "debug", ty = "libc::c_uint", bits = "25..=25")]
+    #[bitfield(name = "trace", ty = "libc::c_uint", bits = "26..=26")]
+    #[bitfield(name = "spare", ty = "libc::c_uint", bits = "27..=27")]
+    #[bitfield(name = "gcgen", ty = "libc::c_uint", bits = "28..=28")]
+    #[bitfield(name = "gccls", ty = "libc::c_uint", bits = "29..=31")]
+    #[bitfield(name = "named", ty = "libc::c_uint", bits = "32..=47")]
+    #[bitfield(name = "extra", ty = "libc::c_uint", bits = "48..=63")]
+    pub rtype_scalar_obj_alt_gp_mark_debug_trace_spare_gcgen_gccls_named_extra: [u8; 8],
+}
+
+pub static mut ACTIVE_BINDING_MASK: libc::c_uint = 1 << 15;
+pub static mut S4_OBJECT_MASK: libc::c_uint = 1 << 4;
+
+impl Sxpinfo {
+
+    pub fn interpret(x: &SEXP) -> &Self {
+        unsafe {
+            (*x as *mut Sxpinfo).as_ref().unwrap()
+        }
+    }
+
+    pub fn is_active(&self) -> bool {
+        self.gp() & unsafe {ACTIVE_BINDING_MASK} != 0
+    }
+
+    pub fn is_immediate(&self) -> bool {
+        self.extra() != 0
+    }
+
+    pub fn is_s4(&self) -> bool {
+        self.gp() & unsafe {S4_OBJECT_MASK} != 0
+    }
+
+    pub fn is_altrep(&self) -> bool {
+        self.alt() != 0
+    }
+
+    pub fn is_object(&self) -> bool {
+        self.obj() != 0
+    }
 }
 
 pub fn r_assert_type(
@@ -75,6 +128,18 @@ pub unsafe fn r_assert_length(
 
 pub fn r_is_null(object: SEXP) -> bool {
     unsafe { object == R_NilValue }
+}
+
+pub fn r_is_altrep(object: SEXP) -> bool {
+    Sxpinfo::interpret(&object).is_altrep()
+}
+
+pub fn r_is_object(object: SEXP) -> bool {
+    Sxpinfo::interpret(&object).is_object()
+}
+
+pub fn r_is_s4(object: SEXP) -> bool {
+    Sxpinfo::interpret(&object).is_s4()
 }
 
 pub fn r_typeof(object: SEXP) -> u32 {
